css part of the login module

+++ app/assets/js/modules/ModalModule.js
@@ -42,20 +42,24 @@ const modalModule = {
   },
 
   // 清除数据库
-  deleteAllTasks() {
-    // 删除数据库
-    db.deleteAllTasks();
+  deleteAllTasks(e) {
+    if (e.target.classList.contains('modal__btn-yes')) {
+      // 删除数据库
+      db.deleteAllTasks();
 
-    // 隐藏模态框
-    const modal = document.querySelector('.modal');
-    modal.classList.remove('modal--visible');
+      // 隐藏模态框
+      const modal = document.querySelector('.modal');
+      modal.classList.remove('modal--visible');
+    }
   },
 
   // 不清除数据库
-  regretDeleteTasks() {
-    // 隐藏模态框
-    const modal = document.querySelector('.modal');
-    modal.classList.remove('modal--visible');
+  regretDeleteTasks(e) {
+    if (e.target.classList.contains('modal__btn-no')) {
+      // 隐藏模态框
+      const modal = document.querySelector('.modal');
+      modal.classList.remove('modal--visible');
+    }
   }
 }
